+++ lib/src/routes/custom_transition.dart
@@ -3,6 +3,8 @@ import 'package:flutter/widgets.dart';
 abstract class CustomTransition {
   Widget buildTransition(
     BuildContext context,
+    Curve curve,
+    Alignment alignment,
     Animation<double> animation,
     Animation<double> secondaryAnimation,
     Widget child,