+++ lib/src/routes/transitions_component.dart
@@ -3,6 +3,8 @@ import 'package:flutter/widgets.dart';
 class TransitionComponent {
   Widget buildChildWithTransition(
       BuildContext context,
+      Curve curve,
+      Alignment alignment,
       Animation<double> animation,
       Animation<double> secondaryAnimation,
       Widget child) {