+++ lib/src/routes/transitions_type.dart
@@ -11,6 +11,8 @@ enum Transition {
   leftToRightWithFade,
   zoom,
   noTransition,
+  cupertino,
+  size,
   native
 }